Remove unecessary demo types, changed mockpaypaldelegate to new file, swift lint, etc
