--- Demo/Demo/Models/BillingAgreementToken.swift
@@ -1,9 +1,0 @@
-import UIKit
-
-struct BillingAgreementToken: Codable {
-
-    enum CodingKeys: String, CodingKey {
-        case tokenID = "token_id"
-    }
-    let tokenID: String
-}
