+++ Demo/Demo/Models/Order.swift
@@ -1,6 +1,5 @@
-// swiftlint:disable space_after_main_type
-
 struct Order: Codable {
+
     let id: String
     let status: String
 }
