+++ Demo/Demo/Models/ProcessOrderParams.swift
@@ -1,6 +1,5 @@
-// swiftlint:disable space_after_main_type
-
 struct ProcessOrderParams: Codable {
+
     let orderId: String
     let intent: String
     let countryCode: String
